feat: Introduce Knowledge Engine and Catalog
- Added a new Knowledge Engine to manage metadata for findings.
- Implemented a Knowledge Catalog to summarize languages, frameworks, runtimes, paradigms, and rules.
- Updated existing modules to utilize the new active knowledge source instead of bundled knowledge.
- Enhanced rendering functions for console and markdown outputs of the knowledge catalog.
- Created workspace scan functionality to handle multiple packages without duplicate findings.
- Added tests to ensure workspace scanning merges packages correctly and maintains unique findings.
- Documented the Knowledge Engine lifecycle and release checklist for version 0.12.0.
