+++ .github/workflows/publish-npm.yml
@@ -4,7 +4,7 @@ on:
   workflow_dispatch:
     inputs:
       tag:
-        description: "Release tag to publish, e.g. v0.11.0. Defaults to package.json version."
+        description: "Release tag to publish, e.g. v0.12.0. Defaults to package.json version."
         required: false
         type: string
   release:
