+++ docs/distribution.md
@@ -113,7 +113,7 @@ Before publishing:
 
 See [docs/release.md](release.md) for the full release process.
 
-For RepoPilot 0.11.0, publish the GitHub Release assets first, then the platform
+For RepoPilot 0.12.0, publish the GitHub Release assets first, then the platform
 npm packages, then the root npm package. The root `repopilot` package declares
 optional dependencies on the platform packages at the same version, so publishing
 the root package before the platform packages can create a broken install window.
@@ -130,7 +130,7 @@ for pkg in \
   @repopilot/win32-x64-msvc; do
   npm view "$pkg" version
 done
-git ls-remote --tags origin v0.11.0
+git ls-remote --tags origin v0.12.0
 ```
 
 ## Homebrew (via tap)
