+++ docs/release.md
@@ -105,11 +105,11 @@ repopilot --version
 brew test repopilot
 ```
 
-For the 0.11.0 release, also verify that the public package and tag state moved
-to 0.11.0:
+For the 0.12.0 release, also verify that the public package and tag state moved
+to 0.12.0:
 
 ```bash
-git ls-remote --tags origin v0.11.0
+git ls-remote --tags origin v0.12.0
 npm view repopilot version
 for pkg in \
   @repopilot/darwin-arm64 \
@@ -119,9 +119,9 @@ for pkg in \
   @repopilot/win32-x64-msvc; do
   npm view "$pkg" version
 done
-npm install -g repopilot@0.11.0
+npm install -g repopilot@0.12.0
 repopilot --version
-repopilot scan . --format json --output /tmp/repopilot-0.11-smoke.json
+repopilot scan . --format json --output /tmp/repopilot-0.12-smoke.json
 ```
 
 If a publishing secret is intentionally absent, publish that channel manually from the exact tagged commit and a clean worktree.
