+++ src/knowledge/catalog/render.rs
@@ -1,196 +1,5 @@
-use crate::knowledge::bundled_knowledge;
-use crate::knowledge::model::SupportLevel;
+use super::model::KnowledgeCatalogReport;
 use crate::output::OutputFormat;
-use serde::Serialize;
-use std::collections::HashSet;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum KnowledgeCatalogSection {
-    All,
-    Languages,
-    Frameworks,
-    Runtimes,
-    Paradigms,
-    Rules,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub struct KnowledgeCatalogReport {
-    pub summary: KnowledgeCatalogSummary,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub languages: Vec<KnowledgeLanguageEntry>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub frameworks: Vec<KnowledgeNamedEntry>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub runtimes: Vec<KnowledgeNamedEntry>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub paradigms: Vec<KnowledgeNamedEntry>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub rules: Vec<KnowledgeRuleEntry>,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub struct KnowledgeCatalogSummary {
-    pub languages: usize,
-    pub frameworks: usize,
-    pub runtimes: usize,
-    pub paradigms: usize,
-    pub rules: usize,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub struct KnowledgeLanguageEntry {
-    pub id: String,
-    pub name: String,
-    pub support: String,
-    pub extensions: Vec<String>,
-    pub filenames: Vec<String>,
-    pub aliases: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub struct KnowledgeNamedEntry {
-    pub id: String,
-    pub name: String,
-}
-
-#[derive(Debug, Clone, Serialize, PartialEq, Eq)]
-pub struct KnowledgeRuleEntry {
-    pub rule_id: String,
-    pub minimum_support: Option<String>,
-    pub languages: Vec<String>,
-    pub frameworks: Vec<String>,
-    pub runtimes: Vec<String>,
-    pub paradigms: Vec<String>,
-    pub suppress_low_signal: bool,
-    pub suppress_generated: bool,
-    pub suppress_config: bool,
-    pub overrides: usize,
-}
-
-pub fn build_knowledge_catalog_report(
-    section: KnowledgeCatalogSection,
-) -> Result<KnowledgeCatalogReport, Box<dyn std::error::Error>> {
-    let knowledge = bundled_knowledge();
-
-    let include_languages = matches!(
-        section,
-        KnowledgeCatalogSection::All | KnowledgeCatalogSection::Languages
-    );
-    let include_frameworks = matches!(
-        section,
-        KnowledgeCatalogSection::All | KnowledgeCatalogSection::Frameworks
-    );
-    let include_runtimes = matches!(
-        section,
-        KnowledgeCatalogSection::All | KnowledgeCatalogSection::Runtimes
-    );
-    let include_paradigms = matches!(
-        section,
-        KnowledgeCatalogSection::All | KnowledgeCatalogSection::Paradigms
-    );
-    let include_rules = matches!(
-        section,
-        KnowledgeCatalogSection::All | KnowledgeCatalogSection::Rules
-    );
-
-    let summary = KnowledgeCatalogSummary {
-        languages: knowledge.languages.len(),
-        frameworks: knowledge.frameworks.len(),
-        runtimes: knowledge.runtimes.len(),
-        paradigms: knowledge.paradigms.len(),
-        rules: knowledge.rule_applicability.len(),
-    };
-
-    let languages = if include_languages {
-        knowledge
-            .languages
-            .iter()
-            .map(|language| KnowledgeLanguageEntry {
-                id: language.id.clone(),
-                name: language.name.clone(),
-                support: support_level_label(language.support).to_string(),
-                extensions: language.extensions.clone(),
-                filenames: language.filenames.clone(),
-                aliases: language.aliases.clone(),
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    let frameworks = if include_frameworks {
-        knowledge
-            .frameworks
-            .iter()
-            .map(|framework| KnowledgeNamedEntry {
-                id: framework.id.clone(),
-                name: framework.name.clone(),
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    let runtimes = if include_runtimes {
-        knowledge
-            .runtimes
-            .iter()
-            .map(|runtime| KnowledgeNamedEntry {
-                id: runtime.id.clone(),
-                name: runtime.name.clone(),
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    let paradigms = if include_paradigms {
-        knowledge
-            .paradigms
-            .iter()
-            .map(|paradigm| KnowledgeNamedEntry {
-                id: paradigm.id.clone(),
-                name: paradigm.name.clone(),
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    let rules = if include_rules {
-        knowledge
-            .rule_applicability
-            .iter()
-            .map(|rule| KnowledgeRuleEntry {
-                rule_id: rule.rule_id.clone(),
-                minimum_support: rule
-                    .minimum_support
-                    .map(support_level_label)
-                    .map(str::to_string),
-                languages: sorted_vec(&rule.languages),
-                frameworks: sorted_vec(&rule.frameworks),
-                runtimes: sorted_vec(&rule.runtimes),
-                paradigms: sorted_vec(&rule.paradigms),
-                suppress_low_signal: rule.suppress_low_signal,
-                suppress_generated: rule.suppress_generated,
-                suppress_config: rule.suppress_config,
-                overrides: rule.overrides.len(),
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    Ok(KnowledgeCatalogReport {
-        summary,
-        languages,
-        frameworks,
-        runtimes,
-        paradigms,
-        rules,
-    })
-}
 
 pub fn render_knowledge_catalog_report(
     report: &KnowledgeCatalogReport,
@@ -373,15 +182,6 @@ fn render_markdown(report: &KnowledgeCatalogReport) -> String {
     output
 }
 
-fn support_level_label(level: SupportLevel) -> &'static str {
-    match level {
-        SupportLevel::DetectOnly => "detect-only",
-        SupportLevel::ImportAware => "import-aware",
-        SupportLevel::ContextAware => "context-aware",
-        SupportLevel::RuleAware => "rule-aware",
-    }
-}
-
 fn comma_or_all(values: &[String]) -> String {
     if values.is_empty() {
         "all".to_string()
@@ -404,10 +204,4 @@ fn markdown_code_list_or_all(values: &[String]) -> String {
     } else {
         markdown_code_list(values)
     }
-}
-
-fn sorted_vec(set: &HashSet<String>) -> Vec<String> {
-    let mut v: Vec<String> = set.iter().cloned().collect();
-    v.sort();
-    v
 }
