+++ src/knowledge/framework.rs
@@ -1,8 +1,8 @@
-use crate::knowledge::bundled_knowledge;
+use crate::knowledge::active_knowledge;
 use crate::knowledge::model::FrameworkProfile;
 
 pub fn profile_by_id(id: &str) -> Option<&'static FrameworkProfile> {
-    bundled_knowledge()
+    active_knowledge()
         .frameworks
         .iter()
         .find(|framework| framework.id == id)
