+++ src/knowledge/paradigm.rs
@@ -1,8 +1,8 @@
-use crate::knowledge::bundled_knowledge;
+use crate::knowledge::active_knowledge;
 use crate::knowledge::model::ParadigmProfile;
 
 pub fn profile_by_id(id: &str) -> Option<&'static ParadigmProfile> {
-    bundled_knowledge()
+    active_knowledge()
         .paradigms
         .iter()
         .find(|paradigm| paradigm.id == id)
