+++ src/knowledge/rule.rs
@@ -1,8 +1,8 @@
-use crate::knowledge::bundled_knowledge;
+use crate::knowledge::active_knowledge;
 use crate::knowledge::model::RuleApplicability;
 
 pub fn applicability_for_rule(rule_id: &str) -> Option<&'static RuleApplicability> {
-    bundled_knowledge()
+    active_knowledge()
         .rule_applicability
         .iter()
         .find(|rule| rule.rule_id == rule_id)
