+++ src/knowledge/runtime.rs
@@ -1,8 +1,8 @@
-use crate::knowledge::bundled_knowledge;
+use crate::knowledge::active_knowledge;
 use crate::knowledge::model::RuntimeProfile;
 
 pub fn profile_by_id(id: &str) -> Option<&'static RuntimeProfile> {
-    bundled_knowledge()
+    active_knowledge()
         .runtimes
         .iter()
         .find(|runtime| runtime.id == id)
